Rename src to cats_and_dogs
