+++ cats_vs_dogs/utils.py
