+++ train.py
@@ -4,8 +4,8 @@ import torchvision
 from nvitop import select_devices
 import torch.nn as nn
 
-from src.runner import Runner
-from src.utils import get_train_val_test_dataloaders
+from cats_vs_docs.runner import Runner
+from cats_vs_docs.utils import get_train_val_test_dataloaders
 
 SIZE_H = SIZE_W = 96
 NUM_CLASSES = 2
